# Add a byte pool for scraping