--- a/pkg/queue/protobuf_stats_reporter.go
+++ b/pkg/queue/protobuf_stats_reporter.go
@@ -18,6 +18,7 @@ package queue
 
 import (
 	"net/http"
+	"strconv"
 	"sync/atomic"
 	"time"
 
@@ -28,7 +29,8 @@ import (
 )
 
 const (
-	contentTypeHeader = "Content-Type"
+	contentTypeHeader   = "Content-Type"
+	contentLengthHeader = "Content-Length"
 )
 
 // PrometheusStatsReporter structure represents a prometheus stats reporter.
@@ -80,6 +82,7 @@ func (r *ProtobufStatsReporter) Handler() http.Handler {
 			return
 		}
 		header.Set(contentTypeHeader, network.ProtoAcceptContent)
+		header.Set(contentLengthHeader, strconv.Itoa(len(buffer)))
 		rsp.Write(buffer)
 	})
 }