Optimizes #8556 as discussed in that PR.

<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

* Adds a multi-size byte slice pool, replacing the `readAll` call which suffers from [certain issues](https://www.openmymind.net/Go-Slices-And-The-Case-Of-The-Missing-Memory/). 
* It sets an upper limit to each slice allocation[ avoiding using a byte buffer](https://github.com/golang/go/issues/23199) and follows [the multi-size pool](https://github.com/golang/go/blob/7e394a2/src/net/http/h2_bundle.go#L998-L1043) approach.
* Benchmarking this with random data shows a significant reduction in memory allocation.
`ReadAll` costs around 750MB while the pool only 18MB.
![normal](https://user-images.githubusercontent.com/7945591/87281340-ca69fb00-c4fb-11ea-8013-d010b3300c27.png)
![opt](https://user-images.githubusercontent.com/7945591/87281348-cccc5500-c4fb-11ea-98d8-6639ec94fe37.png)
```
BenchmarkUnmarshalling-8   	  354242	      3326 ns/op	    3463 B/op	      13 allocs/op  <- current
BenchmarkUnmarshalling-8   	  405806	      2911 ns/op	    1452 B/op	      13 allocs/op  <- optimized
```
Allocation is optimized as shown above.
Ideally we could re-slice buffers to optimize what we add back to the pool but that should not add much.

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```
NONE
```
/assign @markusthoemmes @julz @vagababov