@markusthoemmes 2020-07-13
Can you paste the results of the benchmarks when run via `benchcmp`? That makes them super easy to compare.