@skonto 2020-07-13
@markusthoemmes as i also mentioned to @julz I will do the tests with one size pool and also I will add the `benchcmp` results. The reason I added the profile diagrams above is to show also the memory allocated in total which shows the anticipated issues of `readAll`.