@skonto 2020-07-15
@markusthoemmes @julz @vagababov here are the results:
```
Multiple pool cmp a single with 320 bytes: 

benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      2209          2178          -1.40%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      2280          2277          -0.13%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     2479          2448          -1.25%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      2602          2539          -2.42%

benchmark                                                            old allocs     new allocs     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      19             19             +0.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      20             20             +0.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     22             22             +0.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      21             21             +0.00%

benchmark                                                            old bytes     new bytes     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      1984          1984          +0.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      2064          2064          +0.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     2320          2320          +0.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      2768          2768          +0.00%


Current impl with readAll vs the code in this PR (multiple pools)

benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      2703          2209          -18.28%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      2774          2280          -17.81%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     2998          2479          -17.31%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      3066          2602          -15.13%

benchmark                                                            old allocs     new allocs     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      20             19             -5.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      21             20             -4.76%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     23             22             -4.35%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      22             21             -4.55%

benchmark                                                            old bytes     new bytes     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      4000          1984          -50.40%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      4080          2064          -49.41%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     4336          2320          -46.49%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      4784          2768          -42.14%
```
memory profile of multiple pools:
![mutli](https://user-images.githubusercontent.com/7945591/87541525-0cd13a80-c6aa-11ea-8374-3389009ce619.png)


memory profile of single pool with 320 bytes:

![single](https://user-images.githubusercontent.com/7945591/87541533-1064c180-c6aa-11ea-86e6-478705778881.png)


memory profile of current implementation;

![old](https://user-images.githubusercontent.com/7945591/87541543-13f84880-c6aa-11ea-8c0f-a8eedb05f81b.png)

The diff between a single pool vs multiple is around 10% in these tests where I used a deterministic benchmark with data of different sizes.