@julz 2020-07-15
@skonto could we see current implementation vs. single pool with 320 bytes too? (sorry if I missed it).