@markusthoemmes 2020-07-15
As expected, the single pool is just as quick as the multiple pools but allocates a bit more memory in total. I think the difference is miniscule enough to just go with a single pool and ignore the potential memory savings for now in favor of keeping the code a lot simpler.