@skonto 2020-07-15
@julz the difference is about the memory utilization and it is 10% in my benchmarks:
Here are the numbers for current vs single:

```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      2703          2178          -19.42%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      2774          2277          -17.92%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     2998          2448          -18.35%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      3066          2539          -17.19%

benchmark                                                            old allocs     new allocs     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      20             19             -5.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      21             20             -4.76%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     23             22             -4.35%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      22             21             -4.55%

benchmark                                                            old bytes     new bytes     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      4000          1984          -50.40%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      4080          2064          -49.41%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     4336          2320          -46.49%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      4784          2768          -42.14%
```