@julz 2020-07-15
Yeah, looks like we get 99% of the benefit with much less complexity with a single pool. I vote we go with that.