@skonto 2020-07-15
@markusthoemmes I verified that the client sets `resp.ContentLength` using the following in my benchmark:

```
	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		data := Stat{}
		data.PodName = "a-lzrjc-deployment-85d4b7d859-gspjs"
		data.ProcessUptime = 12.2
		data.ProxiedRequestCount = 122
		data.AverageConcurrentRequests = 2.3
		data.AverageProxiedConcurrentRequests = 100.2
		data.ProxiedRequestCount = 100

		buf, _ := stat.Marshal()
		w.Header().Add("Content-Type", "application/protobuf")
		w.Header().Add("Content-Length", strconv.Itoa(len(buf)))
		if _, err := w.Write(buf); err != nil {
			b.Error(err)
		}
	}))
```
with current code so need to extract the length again.