@vagababov 2020-07-15
If we are going to use a byte buffer, then we can just allocate the largest byte blocks we expect to receive since that's going to graduate there in the end anyway.