@julz 2020-07-15
Yeah, advantage of byte buffer is system will figure out that size on its own, and it avoids allocation in sync.Pool that’ll make staticcheck unhappy (https://staticcheck.io/docs/checks#SA6002)