@skonto 2020-07-16
@julz here are the numbers, single pool with byte slice as it is now vs single pool with a byte buffer of size `PoolSize`:

```
var pool = sync.Pool{
	New: func() interface{} {
		return *bytes.NewBuffer(make([]byte, PoolSize))
	},
}```

```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      2178          2218          +1.84%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      2277          2321          +1.93%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     2448          2506          +2.37%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      2539          2607          +2.68%

benchmark                                                            old allocs     new allocs     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      19             19             +0.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      20             20             +0.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     22             22             +0.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      21             21             +0.00%

benchmark                                                            old bytes     new bytes     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      1984          2000          +0.81%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      2064          2080          +0.78%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     2320          2336          +0.69%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      2768          2784          +0.58%

```

Also here is the mem profile:

![buffer](https://user-images.githubusercontent.com/7945591/87644180-77d74b80-c754-11ea-87a7-cbca3b037e19.png)

There is no significant diff in allocations however with a byte buffer there is almost 40% increase in mem allocation.