@skonto 2020-07-16
@julz @markusthoemmes new numbers comparing only `statFromProto`:

```
benchmark                                                            old ns/op     new ns/op     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      596           134           -77.52%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      644           179           -72.20%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     656           189           -71.19%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      691           230           -66.71%

benchmark                                                            old allocs     new allocs     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      3              1              -66.67%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      4              2              -50.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     4              2              -50.00%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      4              2              -50.00%

benchmark                                                            old bytes     new bytes     delta
BenchmarkUnmarshallingProtoData/BenchmarkStatWithEmptyPodName-8      2096          48            -97.71%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithSmallPodName-8      2144          96            -95.52%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithMediumPodName-8     2192          144           -93.43%
BenchmarkUnmarshallingProtoData/BenchmarkStatWithLargePodName-8      2352          304           -87.07%
```